# docs/api/ezaid-openapi.json
{
  "components": {
    "schemas": {
...
      "AuditPage": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/AuditRecord"
            },
            "title": "Records",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "records"
        ],
        "title": "AuditPage",
        "type": "object"
      },
      "AuditRecord": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "event": {
            "title": "Event",
            "type": "string"
          },
          "generation": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation"
          },
          "request_id": {
            "default": "",
            "title": "Request Id",
            "type": "string"
          },
          "ts": {
            "title": "Ts",
            "type": "string"
          }
        },
        "required": [
          "ts",
          "event",
          "actor"
        ],
        "title": "AuditRecord",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fix": {
            "default": "",
            "title": "Fix",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "serviceToken": {
        "description": "the platform service token (EZAI_CONTROL_TOKEN)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "One authenticated control plane over the platform's declarative state \u2014 Registry v2 + generations, runtime descriptors, governance queue, rendered artifacts \u2014 for the `local-ezai` CLI (connected mode), the Admin Center and the SWE tool server. Every `/v1` call presents the service token as a bearer credential and may forward the human it acts for (`X-EZAI-User`) and its own name (`X-EZAI-Client`); the audit log records `<user> via <client>`. Mutating calls accept an `Idempotency-Key` header: a retry with the same key and payload replays the stored response, a different payload is a conflict. Every error is `{\"error\": {\"code\", \"message\", \"fix\"}}` \u2014 the same object the CLI prints.",
    "title": "Local-EZAI Platform Control Plane (ezaid)",
    "version": "1.0.0-draft.9",
    "x-contract-status": "draft \u2014 frozen by the P2 phase-close PR"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/audit": {
      "get": {
        "operationId": "audit_tail",
        "parameters": [
          {
            "description": "newest records to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "newest records to return",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "forwarded identity of the human the caller acts for",
            "in": "header",
            "name": "X-EZAI-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "forwarded identity of the human the caller acts for",
              "title": "X-Ezai-User"
            }
          },
          {
            "description": "name of the calling surface (cli, admin-center, \u2026)",
            "in": "header",
            "name": "X-EZAI-Client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "name of the calling surface (cli, admin-center, \u2026)",
              "title": "X-Ezai-Client"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "missing or invalid service token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "no audit log attached"
          }
        },
        "security": [
          {
            "serviceToken": []
          }
        ],
        "summary": "Tail of the platform audit log",
        "tags": [
          "governance"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "compose overlay (EZAI_CONTROL_PORT)",
      "url": "http://localhost:8010"
    }
  ]
}